Add MainMenu class file
